requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
curlify==2.2.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic==2.11.4
pydantic_core==2.33.2
PyHamcrest==2.1.0
pytest==8.3.5
requests==2.32.3
structlog==25.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
